squashed image classification into benchmark.py
